chore: add schema layer
